Added C Codegen functions.
These are intended to help generate header files in C for libraries
written in this language. The header files contain opaque struct
definitions for user generated types and prototypes for functions. The
plan is to have the programmer specify which functions to output to the
header file, right now it just does every function in the currently
compiled file.