Begin using the symboltable in the backend.
This cleans up the logic in the backend and reduces the amount of
messing with children of the current node needing to be done. This fixes
an error where inferred variables would crash the compiler. Not
everything is ported yet, and the front end needs to swap over to only
using the symbol table too.