Added new ast node types.
Break and Control are now separate node types. BinOp node type now
indicates if it is a binary or unary operator. These changes help
simplify some type checking and C generation code, letting the switch
statement do the work that used to be done in conditionals. Still one
use of the boolean unaryOp member for BinOp nodes left when the dot
operator is supposed to dereference.